refactor: rename ips_algorithm to algorithm
