+++ src/ips/mod.rs
@@ -1,6 +1,6 @@
 pub mod config;
 mod graph_utils;
-pub mod ips_algorithm;
+pub mod algorithm;
 mod normalization;
 mod peer;
 mod utils;
